Stage 2.4 (section-align Events): rename service + folder + namespace
- IEventGuideService → IEventService
- EventGuideService → EventService
- src/Humans.Application/Interfaces/EventGuide/ → Interfaces/Events/
- src/Humans.Application/Services/EventGuide/ → Services/Events/
- src/Humans.Application/DTOs/EventGuide/ → DTOs/Events/
- namespaces follow

ServiceBoundaryArchitectureTests.RepositoryOwners: IEventRepository
owner section relabelled from "EventGuide" → "Events" so the
CrossSectionRepositoryInjection rule recognises EventService's own repo
as in-section. Baseline file
ApplicationServiceEntityReadReturns.baseline.txt re-namespaced from
Interfaces.EventGuide.IEventGuideService → Interfaces.Events.IEventService.

Test method names that still embed "EventGuide" (separated by
underscores, so no word boundary) get renamed in Stage 2.8 with the
test rearrangement.
