+++ src/Humans.Web/Controllers/CampEventsController.cs
@@ -1,6 +1,6 @@
 using Humans.Application.Interfaces.Camps;
 using Humans.Application.Interfaces.Email;
-using Humans.Application.Interfaces.EventGuide;
+using Humans.Application.Interfaces.Events;
 using Humans.Domain.Entities;
 using Humans.Domain.Enums;
 using Humans.Web.Filters;
@@ -17,7 +17,7 @@ namespace Humans.Web.Controllers;
 [ServiceFilter(typeof(EventGuideFeatureFilter))]
 public class CampEventsController : HumansCampControllerBase
 {
-    private readonly IEventGuideService _guide;
+    private readonly IEventService _guide;
     private readonly IClock _clock;
     private readonly IEmailService _emailService;
     private readonly ILogger<CampEventsController> _logger;
@@ -26,7 +26,7 @@ public class CampEventsController : HumansCampControllerBase
         UserManager<User> userManager,
         ICampService campService,
         IAuthorizationService authorizationService,
-        IEventGuideService guide,
+        IEventService guide,
         IClock clock,
         IEmailService emailService,
         ILogger<CampEventsController> logger)
