+++ tests/Humans.Application.Tests/Architecture/Baselines/ApplicationServiceEntityReadReturns.baseline.txt
@@ -35,28 +35,28 @@ Humans.Application.Interfaces.Dashboard.IDashboardService.GetMemberDashboardAsyn
 Humans.Application.Interfaces.Dashboard.IDashboardService.GetMemberDashboardAsync:Humans.Domain.Entities.Profile
 Humans.Application.Interfaces.Dashboard.IDashboardService.GetMemberDashboardAsync:Humans.Domain.Entities.Shift
 Humans.Application.Interfaces.Dashboard.IDashboardService.GetMemberDashboardAsync:Humans.Domain.Entities.ShiftSignup
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetActiveCategoriesAsync:Humans.Domain.Entities.EventCategory
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetActiveVenuesAsync:Humans.Domain.Entities.EventVenue
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetAllCategoriesAsync:Humans.Domain.Entities.EventCategory
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetAllEventsForDashboardAsync:Humans.Domain.Entities.Event
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetAllVenuesAsync:Humans.Domain.Entities.EventVenue
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetApprovedEventByIdAsync:Humans.Domain.Entities.Event
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetApprovedEventsAsync:Humans.Domain.Entities.Event
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetApprovedEventsForExportAsync:Humans.Domain.Entities.Event
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetApprovedEventsForExportAsync:Humans.Domain.Entities.EventGuideSettings
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetCampEventAsync:Humans.Domain.Entities.Event
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetCampSubmissionsAsync:Humans.Domain.Entities.Event
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetCategoryAsync:Humans.Domain.Entities.EventCategory
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetEventForModerationAsync:Humans.Domain.Entities.Event
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetEventSettingsByIdAsync:Humans.Domain.Entities.EventSettings
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetEventSettingsOptionsAsync:Humans.Domain.Entities.EventSettings
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetEventsByStatusAsync:Humans.Domain.Entities.Event
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetFavouritesWithEventsAsync:Humans.Domain.Entities.EventFavourite
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetGuideSettingsAsync:Humans.Domain.Entities.EventGuideSettings
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetPreferenceAsync:Humans.Domain.Entities.EventPreference
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetUserEventAsync:Humans.Domain.Entities.Event
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetUserSubmissionsAsync:Humans.Domain.Entities.Event
-Humans.Application.Interfaces.EventGuide.IEventGuideService.GetVenueAsync:Humans.Domain.Entities.EventVenue
+Humans.Application.Interfaces.Events.IEventService.GetActiveCategoriesAsync:Humans.Domain.Entities.EventCategory
+Humans.Application.Interfaces.Events.IEventService.GetActiveVenuesAsync:Humans.Domain.Entities.EventVenue
+Humans.Application.Interfaces.Events.IEventService.GetAllCategoriesAsync:Humans.Domain.Entities.EventCategory
+Humans.Application.Interfaces.Events.IEventService.GetAllEventsForDashboardAsync:Humans.Domain.Entities.Event
+Humans.Application.Interfaces.Events.IEventService.GetAllVenuesAsync:Humans.Domain.Entities.EventVenue
+Humans.Application.Interfaces.Events.IEventService.GetApprovedEventByIdAsync:Humans.Domain.Entities.Event
+Humans.Application.Interfaces.Events.IEventService.GetApprovedEventsAsync:Humans.Domain.Entities.Event
+Humans.Application.Interfaces.Events.IEventService.GetApprovedEventsForExportAsync:Humans.Domain.Entities.Event
+Humans.Application.Interfaces.Events.IEventService.GetApprovedEventsForExportAsync:Humans.Domain.Entities.EventGuideSettings
+Humans.Application.Interfaces.Events.IEventService.GetCampEventAsync:Humans.Domain.Entities.Event
+Humans.Application.Interfaces.Events.IEventService.GetCampSubmissionsAsync:Humans.Domain.Entities.Event
+Humans.Application.Interfaces.Events.IEventService.GetCategoryAsync:Humans.Domain.Entities.EventCategory
+Humans.Application.Interfaces.Events.IEventService.GetEventForModerationAsync:Humans.Domain.Entities.Event
+Humans.Application.Interfaces.Events.IEventService.GetEventSettingsByIdAsync:Humans.Domain.Entities.EventSettings
+Humans.Application.Interfaces.Events.IEventService.GetEventSettingsOptionsAsync:Humans.Domain.Entities.EventSettings
+Humans.Application.Interfaces.Events.IEventService.GetEventsByStatusAsync:Humans.Domain.Entities.Event
+Humans.Application.Interfaces.Events.IEventService.GetFavouritesWithEventsAsync:Humans.Domain.Entities.EventFavourite
+Humans.Application.Interfaces.Events.IEventService.GetGuideSettingsAsync:Humans.Domain.Entities.EventGuideSettings
+Humans.Application.Interfaces.Events.IEventService.GetPreferenceAsync:Humans.Domain.Entities.EventPreference
+Humans.Application.Interfaces.Events.IEventService.GetUserEventAsync:Humans.Domain.Entities.Event
+Humans.Application.Interfaces.Events.IEventService.GetUserSubmissionsAsync:Humans.Domain.Entities.Event
+Humans.Application.Interfaces.Events.IEventService.GetVenueAsync:Humans.Domain.Entities.EventVenue
 Humans.Application.Interfaces.GoogleIntegration.IGoogleSyncService.GetRecentOutboxEventsAsync:Humans.Domain.Entities.GoogleSyncOutboxEvent
 Humans.Application.Interfaces.Governance.IApplicationDecisionService.GetApprovedInWindowAsync:Humans.Domain.Entities.Application
 Humans.Application.Interfaces.Governance.IApplicationDecisionService.GetExpiringApplicationsNeedingReminderAsync:Humans.Domain.Entities.Application
