+++ tests/Humans.Application.Tests/Architecture/ServiceBoundaryArchitectureTests.cs
@@ -34,7 +34,7 @@ public class ServiceBoundaryArchitectureTests
             [typeof(IContainerRepository)] = "Containers",
             [typeof(IDriveActivityMonitorRepository)] = "GoogleIntegration",
             [typeof(IEmailOutboxRepository)] = "Email",
-            [typeof(IEventRepository)] = "EventGuide",
+            [typeof(IEventRepository)] = "Events",
             [typeof(IExpenseRepository)] = "Expenses",
             [typeof(IFeedbackRepository)] = "Feedback",
             [typeof(IGeneralAvailabilityRepository)] = "Shifts",
